Corrected the name of the test and added interpolation to the rest of the test dspec
